requirements.txt
mysql-connector-python>=8.0.30
tabulate>=0.8.9
